refactor(Profile): remove logout functionality from Profile component and enhance ProfileInformation with logout button
